Blocks: Update the block's right menu with an open/close toggle
