+++ editor/block-settings-menu/style.scss
@@ -2,14 +2,28 @@
 	position: absolute;
 	top: 0;
 	right: 0;
-	height: $text-editor-font-size * 4; // same height as an empty paragraph
 	padding: 6px 0 6px 14px; // handles hover area
 
 	// Mobile
 	display: none;
 
 	@include break-small {
-		display: block;
+		display: flex;
+		flex-direction: column;
+	}
+}
+
+.editor-block-settings-menu__content {
+	margin-top: 10px;
+}
+
+.editor-block-settings-menu__toggle {
+	transform: rotate(90deg);
+	transition-duration: 0.3s;
+	transition-property: transform;
+
+	&.is-opened {
+		transform: rotate(0);
 	}
 }
 
@@ -35,7 +49,7 @@
 		display: block;
 	}
 
-	&:first-child {
+	.editor-block-settings-menu__content &:first-child {
 		margin-bottom: 4px;
 	}
 }
